Aggregate samples
Aggregates samples into 1 minute windows after a certain interval.

The current aggregation approach avoids the aggregation of time
windows with a single sample.  As a result aggregated time windows are
not considered for aggregation again.

This allows us to avoid a separate flag or model for aggregated
samples. It also allows us to process delayed samples that come in
during the aggregation window using the same logic.